chore: remove obsolete defs
